requirements.txt
django-environ==0.4.5
mysqlclient==2.0.1
requests==2.24.0
